[level 0] Title: 중앙값 구하기, Time: 0.33 ms, Memory: 71.2 MB -BaekjoonHub

+++ 프로그래머스/lv0/120811. 중앙값 구하기/중앙값 구하기.java
@@ -3,7 +3,7 @@ import java.util.Arrays;
 class Solution {
     public int solution(int[] array) {
         Arrays.sort(array);
-        int answer = array[(int)array.length/2];
+        int answer = array[array.length/2];
         return answer;
     }
 }
